refactor: move 'add' command to vue folder
