+++ packages/cli/src/commands/frameworks/index.ts
@@ -1,9 +1,11 @@
 import loadConfigVue from './vue/config'
 import initVue from './vue/init'
+import addVue from './vue/add'
 
 export default {
   vue: {
     loadConfig: loadConfigVue,
     init: initVue,
+    add: addVue,
   },
 }
